pyproject.toml
[project]
name = "spendlyflow"
version = "0.1.0"
description = ""
authors = [
    {name = "SopadeGalinha",email = "jhscial@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=3.2.0, <4.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 79
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.11.13,<0.12.0)"
]
